feat(api): unify AI handlers
